test ready

+++ utils/rename.py
@@ -2,12 +2,14 @@
 import os
 
 os.chdir('../workload')
-for file in os.listdir('.'):
-    wtype, ycsb_op, n_records, n_ops = file.split('-')
-    n_ops = n_ops.split('.')[0]
-    wtype = wtype[-1]
+files = os.listdir('.')
+for file in files:
+    wtype, n_records, n_ops, ycsb_op = file.split('-')
+    n_records = n_records.removesuffix('000000') + 'M'
+    n_ops = n_ops.removesuffix('000000') + 'M'
     new_name = '-'.join([wtype, n_records, n_ops, ycsb_op])
-    if not os.path.exists(new_name):
+    ok = input(f'{file} -> {new_name}. OK? [y / any other key]')
+    if not os.path.exists(new_name) and ok=='y':
         os.rename(file, new_name)
     else:
         raise FileExistsError(f"{new_name} already exists!")
